@@ -29,48 +29,48 @@
     ]
   },
   "dependencies": {
-    "@abtnode/constant": "^1.8.3",
-    "@arcblock/did-auth": "^1.17.5",
-    "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@arcblock/did-connect": "^2.1.63",
-    "@arcblock/ux": "^2.1.63",
-    "@blocklet/sdk": "^1.8.3",
-    "@mui/material": "^5.9.1",
-    "@ocap/client": "^1.17.5",
-    "@ocap/mcrypto": "^1.17.5",
-    "@ocap/wallet": "^1.17.5",
-    "axios": "^0.21.1",
+    "@abtnode/constant": "^1.8.27",
+    "@arcblock/did-auth": "^1.17.22",
+    "@arcblock/did-auth-storage-nedb": "^1.7.1",
+    "@arcblock/did-connect": "^2.4.36",
+    "@arcblock/ux": "^2.4.36",
+    "@blocklet/sdk": "^1.8.27",
+    "@mui/material": "^5.10.8",
+    "@ocap/client": "^1.17.22",
+    "@ocap/mcrypto": "^1.17.22",
+    "@ocap/wallet": "^1.17.22",
+    "axios": "^0.21.4",
     "compression": "^1.7.4",
-    "cookie-parser": "^1.4.5",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv-flow": "^3.2.0",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
     "express-history-api-fallback": "^2.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "4.0.3",
-    "styled-components": "^5.3.5"
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
     "@arcblock/eslint-config": "latest",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^13.5.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.31.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-prettier": "^3.4.1",
+    "eslint-plugin-react": "^7.31.8",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-unicorn": "^34.0.1",
-    "http-proxy-middleware": "^2.0.1",
-    "nodemon": "^2.0.12",
+    "http-proxy-middleware": "^2.0.6",
+    "nodemon": "^2.0.20",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.2"
+    "prettier": "^2.7.1"
   }
 }
